refactor: Add code to repeat async image queries
When the code inside the callback function of a then() statement
finishes execution, the function surrounding the then() statement
is not exited, nor is the surrounding function affected by any
statements within the then() statement's callback function.
Therefore, nested then() statements are not an effective way
of performing repeated calls to an asynchronous function,
especially if the repetition has a condition based on the
data returned from each call.

Instead of trying to perform multiple image queries by calling
multiple then() statements inside a loop on the server-side
code (which may cause an error if multiple responses are sent,
due to the response headers being set after one response is sent),
use a separate asynchronous function on the client-side code.
This asynchronous function uses the 'async' and 'await' keywords
to easily extract data contained inside Promises. Therefore, nested
then() statements are not needed to perform repeated calls
to an asynchronous function.