fix: Remove diacritics from image query
Sometimes, fetch calls to the Pixabay API will return an error
if diacritics are included in the query. This is due to some
security restrictions implemented by the Pixabay API's security
service. Therefore, diacritics must be removed from the query
before using it in a fetch call to the Pixabay API.