refactor: Add button to show list of trips
Instead of loading the list of scheduled trips as the page loads,
provide a button that the user can use to load the list.
In the case that the Express server has not been started before the
webpage requests the list of trips, the webpage would then
generate an error message on the user interface. Therefore, loading
the trip list as the page loads might cause this error. This might have
confused the user because the page would have barely been loaded.
Therefore, a button is added so that the user can control when the
list is loaded, and so that the aforementioned error message is shown
after the user interacts with the page. Thus, the user will not be
confused about what caused the aforementioned error message.